+++ packages/nodes-base/package.json
@@ -447,7 +447,6 @@
       "dist/nodes/Hubspot/HubspotTrigger.node.js",
       "dist/nodes/HumanticAI/HumanticAi.node.js",
       "dist/nodes/Hunter/Hunter.node.js",
-      "dist/nodes/FileSystemWatch.node.js",
       "dist/nodes/ICalendar.node.js",
       "dist/nodes/If.node.js",
       "dist/nodes/Iterable/Iterable.node.js",
@@ -469,6 +468,7 @@
       "dist/nodes/Line/Line.node.js",
       "dist/nodes/LingvaNex/LingvaNex.node.js",
       "dist/nodes/LinkedIn/LinkedIn.node.js",
+      "dist/nodes/LocalFileTrigger.node.js",
       "dist/nodes/Magento/Magento2.node.js",
       "dist/nodes/MailerLite/MailerLite.node.js",
       "dist/nodes/MailerLite/MailerLiteTrigger.node.js",
